Politica para posts
He creado una politica para que nadie pueda editar ni borrar posts ajenos, anteriormente solo habia ocultado los botones si no coincidia el id pero se podia acceder desde la ruta, asi que con la politica ya queda cubierto, de momento solamente las opciones de actualizar y borrar, seguramente la politica crezca mas en el futuro.

+++ app/Http/Controllers/PostController.php
@@ -8,9 +8,12 @@ use Barryvdh\DomPDF\Facade\Pdf;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class PostController extends Controller
 {
+    use AuthorizesRequests;
+
     public function index()
     {
 
@@ -49,6 +52,7 @@ class PostController extends Controller
     public function edit(Post $post)
     {
 
+        $this->authorize('update', $post);
         return view('edit', compact('post'));
 
     }
@@ -68,8 +72,10 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
-        $post->delete();
 
+        $this->authorize('delete', $post);
+
+        $post->delete();
         return redirect('home');
 
     }
